requirements.txt
cat > requirements.txt << 'EOF'
fastapi
uvicorn
gitpython
openai
EOF